# docs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RobotOps Sentinel API",
    "summary": "Closed-loop robotics supervision backend.",
    "license": {
      "name": "Proprietary hackathon prototype",
      "identifier": "LicenseRef-Proprietary"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local development backend"
    }
  ],
  "paths": {
...
    "/api/perception/detect": {
      "post": {
        "tags": [
          "perception"
        ],
        "summary": "Detect Incident Objects",
        "description": "Detect incident-relevant objects in a camera frame.",
        "operationId": "detect_incident_objects_api_perception_detect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeminiObjectDetectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeminiObjectDetectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/perception/semantic-status": {
      "post": {
        "tags": [
          "perception"
        ],
        "summary": "Describe Semantic Status",
        "description": "Describe the semantic status of a camera frame.",
        "operationId": "describe_semantic_status_api_perception_semantic_status_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeminiObjectDetectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SemanticStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GeminiDetectionBox": {
        "properties": {
          "x_min": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "X Min"
          },
          "y_min": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Y Min"
          },
          "x_max": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "X Max"
          },
          "y_max": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Y Max"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "x_min",
          "y_min",
          "x_max",
          "y_max"
        ],
        "title": "GeminiDetectionBox",
        "description": "Normalized detection box in image coordinates."
      },
      "GeminiObjectDetection": {
        "properties": {
          "label": {
            "type": "string",
            "minLength": 1,
            "title": "Label"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "box": {
            "$ref": "#/components/schemas/GeminiDetectionBox"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "label",
          "box"
        ],
        "title": "GeminiObjectDetection",
        "description": "Incident-relevant object detected by Gemini Robotics-ER."
      },
      "GeminiObjectDetectionRequest": {
        "properties": {
          "image_base64": {
            "type": "string",
            "maxLength": 1500000,
            "minLength": 1,
            "title": "Image Base64"
          },
          "mime_type": {
            "$ref": "#/components/schemas/InlineImageMimeType",
            "default": "image/jpeg"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "image_base64"
        ],
        "title": "GeminiObjectDetectionRequest",
        "description": "Inline camera frame for Gemini Robotics-ER object detection."
      },
      "GeminiObjectDetectionResponse": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model"
          },
          "detected_at": {
            "type": "string",
            "format": "date-time",
            "title": "Detected At"
          },
          "detections": {
            "items": {
              "$ref": "#/components/schemas/GeminiObjectDetection"
            },
            "type": "array",
            "title": "Detections"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "model",
          "detected_at"
        ],
        "title": "GeminiObjectDetectionResponse",
        "description": "Gemini Robotics-ER object detections for a camera frame."
      },
      "GeminiSemanticStatusResult": {
        "properties": {
          "state": {
            "$ref": "#/components/schemas/PathState"
          },
          "blocking_object": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blocking Object"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "rationale": {
            "type": "string",
            "minLength": 1,
            "title": "Rationale"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "state",
          "confidence",
          "rationale"
        ],
        "title": "GeminiSemanticStatusResult",
        "description": "Structured semantic status for the visible path."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InlineImageMimeType": {
        "type": "string",
        "enum": [
          "image/jpeg",
          "image/png",
          "image/webp"
        ]
      },
...
      "PathState": {
        "type": "string",
        "enum": [
          "clear",
          "blocked"
        ],
        "title": "PathState",
        "description": "Semantic state of the visible path."
      },
...
      "SemanticStatusResponse": {
        "properties": {
          "assessed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Assessed At"
          },
          "status": {
            "$ref": "#/components/schemas/GeminiSemanticStatusResult"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "assessed_at",
          "status"
        ],
        "title": "SemanticStatusResponse",
        "description": "Semantic status result for a camera frame."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}